order-status-service/requirements.txt
fastapi
uvicorn
psycopg2-binary
SQLAlchemy
bcrypt
PyJWT